changed first x variable to double data type. output become a double data type aswell by doing so.

+++ Activity8CodeCorrection.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main() {
 
-    int x = 10;
+    double x = 10;
 
     int y = 3;
